refactor: update va-input-wrapper imports/exports
