+++ packages/ui/src/components/va-input/VaInput.vue
@@ -87,7 +87,7 @@ import { useCleave, useCleaveProps } from './hooks/useCleave'
 
 import type { AnyStringPropType } from '../../types/prop-type'
 
-import VaInputWrapper from './components/VaInputWrapper.vue'
+import VaInputWrapper from './components/VaInputWrapper/VaInputWrapper.vue'
 import VaTextarea from './components/VaTextarea/VaTextarea.vue'
 import VaIcon from '../va-icon/VaIcon.vue'
